RaspberryPi support
Thanks a lot to zatrax for helping me in converting docbender's Onkyo-RI to using WiringPi instead of the Arduino library. 

A makefile has also been created for compiling the code with g++. Please compile on your Raspberry pi using the makefile (modify to your taste). Please also modify the main.cpp to the code you want to send.